% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{acmb}
\alias{acmb}
\title{This function shows absolute counts of each cluster per each sample(or group)}
\usage{
acmb(
  sobj,
  identity = "seurat_clusters",
  group.by = "sample",
  idents = NULL,
  df = F,
  vlines = NULL,
  vline_color = "red"
)
}
\description{
This function shows absolute counts of each cluster per each sample(or group)
}
