% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{add_signature_enrichit}
\alias{add_signature_enrichit}
\title{Add Gene Signature Score using enrichIt (Improved Version)}
\usage{
add_signature_enrichit(
  seurat_obj,
  gene_source,
  signature_name,
  input_keytype = "ENSEMBL",
  gene_col = 1,
  sheet_name = 1,
  assay = "RNA",
  layer = "data",
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{gene_source}{A character string specifying the path to a file OR an R object (data.frame, vector) containing gene IDs.}

\item{signature_name}{A character string for the name of the new metadata column.}

\item{input_keytype}{The type of the input gene IDs. Must be a valid keytype for org.Hs.eg.db.
Common values are "ENSEMBL", "SYMBOL", "ENTREZID". Defaults to "ENSEMBL".}

\item{gene_col}{If \code{gene_source} is a file path or a data.frame, specify the column index or name. Defaults to 1.}

\item{sheet_name}{If \code{gene_source} is an xlsx file, specify the sheet name or index. Defaults to 1.}

\item{assay}{The assay to use from the Seurat object. Defaults to "RNA".}

\item{layer}{The layer (slot) to use from the assay. Defaults to "data".}

\item{...}{Additional arguments to be passed to \code{escape::enrichIt}.}
}
\value{
A Seurat object with the new signature score added to its metadata.
}
\description{
Calculates a gene signature score using escape::enrichIt and adds it to the Seurat object's metadata.
This version flexibly accepts different gene ID types (e.g., ENSEMBL, SYMBOL, ENTREZID).
}
\examples{
\dontrun{
# Excel 파일에서 ENSEMBL ID를 읽어서 InflammatoryScore 시그니처 점수를 추가하는 예시
data_seurat <- add_signature_enrichit(
  seurat_obj = data_seurat,
  gene_source = "/data/kjc1/mylit/projects/mIBD/41590_2024_1994_MOESM6_ESM.xlsx",
  signature_name = "InflammatoryScore",
  input_keytype = "ENSEMBL",
  gene_col = 1,
  sheet_name = "Inflammation_score",
  assay = "RNA",
  layer = "data"
)
}
}
