% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_frequency.R
\name{add_simple_pvalues}
\alias{add_simple_pvalues}
\title{Add p-values without brackets (similar to first image style)}
\usage{
add_simple_pvalues(plot_obj, stats_df, plot_data, p_label_size = 3)
}
\arguments{
\item{plot_obj}{ggplot object}

\item{stats_df}{Statistics dataframe}

\item{plot_data}{Plot data with x_labels}

\item{p_label_size}{Size of p-value text}
}
\value{
Modified ggplot object
}
\description{
Add p-values without brackets (similar to first image style)
}
