% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{AddMultipleModuleScores}
\alias{AddMultipleModuleScores}
\title{Add Module Scores for a List of Feature Sets}
\usage{
AddMultipleModuleScores(
  seurat_object,
  feature_sets,
  assay = NULL,
  slot = "data",
  nbin = 24,
  ctrl = 100,
  seed = 1,
  search = FALSE,
  ...
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{feature_sets}{A named or unnamed list of character vectors. Each character vector is a set of gene names.
If the list is named, the names will be used for the output module score columns.
If a list element is unnamed, its name will be generated by concatenating gene names with "+".}

\item{assay}{Name of the assay to use. Default is the current default assay.}

\item{slot}{Slot to pull expression data from. Default is "data".}

\item{nbin}{Number of bins for AddModuleScore. Default is 24.}

\item{ctrl}{Number of control features for AddModuleScore. Default is 100.}

\item{seed}{Random seed for AddModuleScore. Default is 1.}

\item{search}{Passed to Seurat::\verb{[.Assay}. Default is FALSE. Relevant if features are not present in the object.}

\item{...}{Additional arguments passed to Seurat::AddModuleScore.}
}
\value{
A Seurat object with added module scores in the metadata.
}
\description{
This function takes a Seurat object and a list of feature (gene) sets.
For each feature set, it calculates a module score using Seurat's AddModuleScore.
The name of the resulting metadata column will be the name of the feature set if provided,
or a concatenation of gene names joined by "+" if the feature set is unnamed.
}
\examples{
\dontrun{
# Assuming 'pbmc' is a Seurat object with an RNA assay
# Example feature sets
gene_sets <- list(
  Tcell_activation = c("CD69", "IFNG", "TNF"),
  Monocyte_markers = c("CD14", "LYZ"),
  c("CCR7", "SELL", "LEF1"), # Unnamed set
  Bcell_core = c("MS4A1", "CD19", "CD79A")
)

pbmc <- AddMultipleModuleScores(seurat_object = pbmc, feature_sets = gene_sets)

# Check the new metadata columns
head(pbmc@meta.data)

# Visualize a score
FeaturePlot(pbmc, features = "Tcell_activation1") # Name is appended with a number
FeaturePlot(pbmc, features = "CCR7_SELL_LEF11")
}
}
