% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agi.plot.R
\name{agi.plot}
\alias{agi.plot}
\title{Gera os graficos de AGI}
\usage{
agi.plot(dados, desemp, redGab = NULL, main = "AGI PLOT",
  files = NULL, xlab = "Escore", Alts = c(".", "*", "A", "B", "C",
  "D", "E"), col.agi = NULL, cex = c(3, 1.2, rep(0.7, 5)),
  density = TRUE, smooth = NULL, dir.create = "AGI", xlim = NULL,
  width = 9.6, height = 6.8, shinyDemo = NULL)
}
\arguments{
\item{dados}{Banco de dados a ser utilizado. Este deve ser composto com os seguintes campos: (IT1, ..., ITm).}

\item{desemp}{Medida de desempenho a ser utilizada.}

\item{redGab}{Vetor de gabaritos. Caso este argumento seja passado a funcao, as curvas referentes as marcacoes corretas serao printadas com a cor vermelha.}

\item{main}{String ou vetor com o(s) titulo(s) dos graficos.}

\item{files}{Vetor com o nome dos arquivos. Caso NULL, serao utilizados os nomes dos itens no objeto 'dados'.}

\item{xlab}{Label do eixo x.}

\item{Alts}{Vetor com as possiveis marcacoes presentes nos dados.}

\item{col.agi}{Vetor com as coloracoes a serem utilizadas nas curvas. A ordem das cores sera equivalente a ordem no objeto 'Alts'. Caso redGab, a coloracao do gabarito sera ignorada. Caso NULL, coloracoes aleatorias serao utilizadas.}

\item{cex}{Vetor com as dimensoes dos plots de 'Alts' perante as curvas. A ordem dos elementos sera equivalente a ordem de 'Alts' e 'col'.}

\item{density}{Plotar na imagem a distribuicao dos escores.}

\item{smooth}{Suavizar as curvas de comportamento, interpolando 'smooth' pontos na curva.}

\item{dir.create}{Nome do diretorio a ser criado para guardar os graficos.}

\item{xlim}{Limites do eixo X a ser plotados.}

\item{width}{Largura em polegadas da imagem.}

\item{height}{Altura em polegadas da imagem.}

\item{shinyDemo}{Argumento auxiliar para o uso da funcao 'myR.app'.}
}
\description{
Cria um diretorio chamado 'AGI' e gera os graficos de Analise Grafica do Item (AGI).
}
\details{
Etapa Anterior: 'escore'.
}
\examples{
bib = list(CAD1 = c(1,2), CAD2 = c(2,3), CAD3 = c(3,1))

nblocos = 3

b.len = 2

resp = data.frame(matrix(sample(c("A", "B", "C"), 40, replace = T), ncol = 4), stringsAsFactors = F); names(resp) = paste0("IT", 1:4)

dados = cbind(data.frame(ID = 1:10, CAD = paste0("CAD", sample(3, 10, replace = T)), PESO = 1), resp)

dados = vetor.extendido(dados, bib, nblocos, b.len)

dadosCorr = corrigir(dados, c("A", "B", "A", "C", "A", "B"))

dadosCorr = escore(dadosCorr)

agi.plot(dados[-c(1:3, ncol(dados))], desemp = dadosCorr$Escore, redGab = c("A", "B", "A", "C", "A", "B"), main = paste0("AGI ", names(dados)[-c(1:3, ncol(dados))]), Alts = c("A", "B", "C"))

}
\author{
Brunno Bittencourt
}
