% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.plot.R
\name{all.plot}
\alias{all.plot}
\title{Gera os graficos de AGI, CCI e DIF}
\usage{
\method{all}{plot}(dados, desemp, pars, EXP, SCO, xlab = c("Escore",
  "Proficiencia", "Proficiencia"), ylab = c("Percentual de Respostas",
  "Probabilidade", "Percentual de Respostas"), main = "ALL PLOT",
  files = NULL, redGab = NULL, xlim.agi = NULL, smooth = NULL,
  Alts = c(".", "*", "A", "B", "C", "D", "E"), col.agi = NULL,
  cex = c(3, 1.2, rep(0.7, 5)), breaks = c(-4, 4, 0.1),
  groups = NULL, probs = c(0.05, 0.95), col.dif = NULL,
  xlim.dif = c(-4, 4), dir.create = "Full Plot", width = 9.6,
  height = 6.8, shinyDemo = NULL)
}
\arguments{
\item{dados}{Banco de dados a ser utilizado na AGI. Este deve ser composto com os seguintes campos: (IT1, ..., ITm).}

\item{desemp}{Medida de desempenho a ser utilizada na AGI.}

\item{pars}{DF com os parametros dos itens a serem utilizados para a CCI. Deve ser composto com os seguintes campos: (Codigo, a, b, c).}

\item{EXP}{Objeto resultante da funcao calcDIF.}

\item{SCO}{Vetor de proficiencias para o plot da densidade.dddddddddddddddd}

\item{xlab}{Vetor de labels para os eixos x dos tres graficos.}

\item{ylab}{Vetor de labels para os eixos y dos tres graficos.}

\item{main}{String ou vetor com o(s) titulo(s) dos graficos.}

\item{files}{Vetor com o nome dos arquivos. Caso NULL, serao utilizados os nomes dos itens no objeto 'pars'.}

\item{redGab}{Vetor de gabaritos. Caso este argumento seja passado a funcao, as curvas referentes as marcacoes corretas na AGI serao printadas com a cor vermelha.}

\item{xlim.agi}{Limites do eixo X a ser plotados na AGI.}

\item{smooth}{Suavizar as curvas da AGI interpolando a curva atraves de 'nCat * smooth' pontos na curva, sendo nCat o numero de possiveis escores.}

\item{Alts}{Vetor com as possiveis marcacoes presentes nos dados.}

\item{col.agi}{Vetor com as coloracoes a serem utilizadas nas curvas. A ordem das cores sera equivalente a ordem no objeto 'Alts'. Caso redGab, a coloracao do gabarito sera ignorada. Caso NULL, coloracoes aleatorias serao utilizadas.}

\item{cex}{Vetor com as dimensoes dos plots de 'Alts' perante as curvas da AGI. A ordem dos elementos sera equivalente a ordem de 'Alts' e 'col.agi'.}

\item{breaks}{Limite inferior e superior dos pontos de quadratura na CCI, acompanhados pelo step da sequencia.}

\item{groups}{Vetor com o nome a ser plotado para cada um dos grupos nos graficos de DIF, ordenado pelo indice do respectivo grupo.}

\item{probs}{Quantiles utilizados para o calculo dos intervalos de interesse na analise de DIF.}

\item{col.dif}{Vetor com as cores a serem utilizadas nas curvas empiricas dos grupos nos graficos de DIF.}

\item{xlim.dif}{Limites do eixo X a ser plotados nos graficos de DIF.}

\item{dir.create}{Nome do diretorio a ser criado para guardar os graficos.}

\item{width}{Largura em polegadas da imagem.}

\item{height}{Altura em polegadas da imagem.}

\item{shinyDemo}{Argumento auxiliar para o uso da funcao 'myR.app'.}
}
\description{
Cria um diretorio chamado 'ALL PLOT' e gera os graficos dos item conjuntamente.
}
\details{
Etapa Anterior: 'escore', 'checkDIF', 'read.SCO'.

Os objetos 'dados', 'pars' e 'EXP' devem estar na mesma ordenacao em relacao aos itens. Isto e, a quinta coluna de 'dados' deve ser referente a quinta linha de 'pars', que por sua vez e equivalente ao quinto elemento de 'EXP'.
}
\examples{
bib = list(CAD1 = c(1,2), CAD2 = c(2,3), CAD3 = c(3,1))

nblocos = 3

b.len = 2

resp = data.frame(matrix(sample(c("A", "B", "C"), 40, replace = T), ncol = 4), stringsAsFactors = F); names(resp) = paste0("IT", 1:4)

dados = cbind(data.frame(ID = 1:10, CAD = paste0("CAD", sample(3, 10, replace = T)), PESO = 1), resp)

dados = vetor.extendido(dados, bib, nblocos, b.len)

dadosCorr = corrigir(dados, c("A", "B", "A", "C", "A", "B"))

dadosCorr = escore(dadosCorr)

pars = data.frame(Codigo = 1001:1006, a = abs(rnorm(6, mean = 1.5, sd = 0.5)), b = rnorm(6, mean = 0.5), c = runif(6, 0.1, 0.3))

EXP = remakeEXP(readLines("EXPtest.EXP"))

newGroup = c(6, 7)

EXP = calcDIF(EXP, newGroup)

SCO = read.SCO(readLines("SCOtest.SCO"))

#Selecionando itens do teste.
EXP = EXP[names(EXP) \%in\% c("IT23094", "IT23095", "IT20345", "IT23094", "IT34909", "IT23094")

all.plot(dados[-c(1:3, ncol(dados))], desemp = dadosCorr$Escore, redGab = c("A", "B", "A", "C", "A", "B"), xlim.agi = c(0, 6), smooth = 5, pars = pars, SCO = SCO, EXP = EXP)

}
\author{
Brunno Bittencourt
}
