% Generated by roxygen2 (4.0.2): do not edit by hand
\name{areaUnderCurve}
\alias{areaUnderCurve}
\title{Area Under The Curve (AUC)}
\usage{
areaUnderCurve(y, probs, plot = FALSE)
}
\arguments{
\item{y}{a vector of ones and zeros descirbing presense/absense}

\item{probs}{predicted probabilities}

\item{plot}{logical value indicating if the ROC should be ploted.}
}
\value{
a scalar value
}
\description{
Compute the area under the receiver operating curve (AUC) based on the trapezoid method.
}
\details{
See s. 38-42 in reference and Rune's excel ark.
}
\author{
Hans Ole &Oslashrka \email{hans.ole.orka@gmail.org}
}
\references{
Halvorsen, R. 2012. A maximum likelihood explanation of MaxEnt, and some
implications for distribution modelling. Sommerfeltia 36: 1-XXX. Oslo. ISBN 82-7420-0XXX.
ISSN 0800-6865.
}

