% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buteo_age.R
\name{buteo_age}
\alias{buteo_age}
\title{Estimate the age of juvenile buzzards based on wing length}
\usage{
buteo_age(df = NULL, wing = "wing", sex = "sex", unit = c("mm",
  "cm"), .plot = F)
}
\arguments{
\item{df}{data frame containing data}

\item{wing}{column name for wing measurements}

\item{sex}{column name giving the sex of individuals or NULL}

\item{unit}{unit of values. By default in mm}
}
\description{
Buzzard age is estimated based on a standard growth curve modeled from published data by Rob G. Bijlsma (1999)
}
\examples{
## 240 mm wing length
buteo_age(data.frame(wing = c(240,240), sex = c("male","female")))

}
\references{
Bijlsma, RG 1999: Sex determination of nestling Common Buzzards Buteo buteo. Limosa 72, 1.10
}
