% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buteo_condition.R
\name{buteo_condition}
\alias{buteo_condition}
\title{residual weight with respect to a regression of wing length against weight}
\usage{
buteo_condition(df = NULL, wing = "wing", weight = "weight",
  sex = "sex", unit = c("cm", "mm"), .plot = F)
}
\arguments{
\item{df}{data frame containing data}

\item{wing}{name of variable giving wing length}

\item{weight}{name of variable giving weight [g]}

\item{sex}{name of variable giving sex}

\item{unit}{unit of wing measurements}
}
\description{
residual weight with respect to a regression of wing length against weight
}
\examples{
# NOT RUN
# buteo_condition(data.frame(wing = c(240,240), sex = c(0,1), weight = c(700,700)))

}
\references{
Bijlsma, RG 1999: Sex determination of nestling Common Buzzards Buteo buteo. Limosa 72, 1.10
}
