% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDIF.R
\name{calcDIF}
\alias{calcDIF}
\title{Calcula as diferencas entre as proporcoes}
\usage{
calcDIF(EXP, newGroup)
}
\arguments{
\item{EXP}{Objeto resultante da funcao remakeEXP.}

\item{newGroup}{Vetor com os grupos de interesse (so serao plotados os itens com pelo menos um grupo de interesse).}
}
\value{
Lista com um DF reestruturado para cada item do arquivo.
}
\description{
Para cada DF dentro da lista resultante da funcao 'remakeEXP', calcula as diferencas entre as PROPORTIONs dos grupos de interesse.
}
\details{
Etapa Anterior: 'remakeEXP'.

Etapa Posterior: 'checkDIF'.
}
\examples{
EXP = remakeEXP(readLines("EXPtest.EXP"))

newGroup = c(6, 7)

EXP = calcDIF(EXP, newGroup)

}
\author{
Brunno Bittencourt
}
