% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_utils.R
\name{calculate_ribosomal_percentage}
\alias{calculate_ribosomal_percentage}
\title{Calculate Ribosomal Gene Percentage}
\usage{
calculate_ribosomal_percentage(
  obj,
  rps_pattern = "^RPS",
  rpl_pattern = "^RPL",
  assay = "RNA"
)
}
\arguments{
\item{obj}{Seurat object}

\item{rps_pattern}{Pattern for RPS genes (default: "^RPS")}

\item{rpl_pattern}{Pattern for RPL genes (default: "^RPL")}

\item{assay}{Assay to use (default: "RNA")}
}
\value{
Seurat object with percent.ribo added to metadata
}
\description{
Calculates percentage of ribosomal genes, excluding known non-ribosomal genes
}
