% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.plot.R
\name{cci.plot}
\alias{cci.plot}
\title{Gera os graficos da CCI}
\usage{
cci.plot(pars, main = "CCI PLOT", SCO = NULL, breaks = c(-4, 4, 0.1),
  files = NULL, dir.create = "CCI", width = 9.6, height = 6.8,
  shinyDemo = NULL)
}
\arguments{
\item{pars}{DF com os parametros dos itens. Deve ser composto com os seguintes campos: (Codigo, a, b, c).}

\item{main}{String ou vetor com o(s) titulo(s) dos graficos.}

\item{SCO}{Vetor de proficiencias para o plot da densidade.}

\item{breaks}{Limite inferior e superior dos pontos de quadratura, acompanhados pelo step da sequencia.}

\item{files}{Vetor com o nome dos arquivos. Caso NULL, serao utilizados os nomes dos itens no objeto 'dados'.}

\item{dir.create}{Nome do diretorio a ser criado para guardar os graficos.}

\item{width}{Largura em polegadas da imagem.}

\item{height}{Altura em polegadas da imagem.}

\item{shinyDemo}{Argumento auxiliar para o uso da funcao 'myR.app'.}
}
\description{
Cria um diretorio chamado 'CCI' e gera os graficos da Curva Caracteristica do Item (CCI).
}
\details{
Etapa Anterior: 'read.SCO'.
}
\examples{
pars = data.frame(Codigo = 1001:1006, a = abs(rnorm(6, mean = 1.5, sd = 0.5)), b = rnorm(6, mean = 0.5), c = runif(6, 0.1, 0.3))
			  
cci.plot(pars, paste0("Item ", pars$Codigo))

}
\author{
Brunno Bittencourt
}
