% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDIF2.R
\name{checkDIF2}
\alias{checkDIF2}
\title{Checa os itens com DIF e mal ajuste}
\usage{
checkDIF2(EXP, SCO, newGroup, crit = 0.15, minresp = 20,
  probs = c(0.05, 0.95))
}
\arguments{
\item{EXP}{Objeto resultante da funcao calcDIF.}

\item{SCO}{DF com o Grupo e o SCO de cada individuo.}

\item{newGroup}{Vetor com os grupos de interesse (so serao plotados os itens com pelo menos um grupo de interesse).}

\item{crit}{Diferenca minima para caracterizar um comportamento diferencial.}

\item{minresp}{Numero minimo de respostas para utilizacao do ponto de quadratura em questao.}

\item{probs}{Quantiles utilizados para o calculo dos intervalos de interesse.}
}
\value{
Lista com dois DFs, apresentando os itens que apresentaram comportamento diferencial e mal ajuste.
}
\description{
Funcao em desenvolvimento. Implementacao do argumento minresp.
}
\details{
Apenas serao consideradas as proporcoes entre os quantis passados no argumento SCO.

Etapa Anterior: 'calcDIF' 'read.SCO'.

Etapa Posterior: 'dif.plot'.
}
\examples{
EXP = remakeEXP(readLines("EXPtest.EXP"))

newGroup = c(6, 7)

EXP = calcDIF(EXP, newGroup)

SCO = read.SCO(readLines("SCOtest.SCO"))

SCO = lapply(split(SCO$SCO, SCO$Grupo), quantile, probs = c(.05, .95))

DIF = checkDIF(EXP, SCO, newGroup)

}
\author{
Brunno Bittencourt
}
