% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_dim_reduction.R
\name{clr_transform}
\alias{clr_transform}
\title{Closed Log-Ratio Transform}
\usage{
clr_transform(X, eps = 1e-06)
}
\arguments{
\item{X}{Numeric matrix (rows = samples, columns = components).}

\item{eps}{Pseudo-count added before log transform.}
}
\value{
Matrix on CLR scale.
}
\description{
Apply a CLR transform (with pseudo-count) to compositional matrices.
}
