% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{CompareModuleScoringMethods}
\alias{CompareModuleScoringMethods}
\title{Compare Module Scoring Methods}
\usage{
CompareModuleScoringMethods(
  sobj,
  gene_sets,
  group = "seurat_clusters",
  assay = "SCT"
)
}
\description{
This function compares the simple averaging method with AddModuleScore
}
