% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{corr_pairwise}
\alias{corr_pairwise}
\title{Pairwise correlation among a set of features.
Optionally grouped or split (same semantics as corr_with_major).}
\usage{
corr_pairwise(
  data,
  features,
  method = c("pearson", "spearman"),
  group.by = NULL,
  split.by = NULL,
  assay = "RNA",
  slot = "scale.data"
)
}
\value{
A list of correlation matrices (if split) or a single tibble of pairs.
}
\description{
Pairwise correlation among a set of features.
Optionally grouped or split (same semantics as corr_with_major).
}
