% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{corr_with_major}
\alias{corr_with_major}
\title{Compute correlation (Pearson/Spearman) of each feature with a major feature.
Handles optional grouping & splitting.}
\usage{
corr_with_major(
  data,
  major_feature,
  features,
  method = c("pearson", "spearman"),
  group.by = NULL,
  split.by = NULL,
  assay = "RNA",
  slot = "scale.data"
)
}
\value{
A list (if split.by provided) or tibble.
}
\description{
Compute correlation (Pearson/Spearman) of each feature with a major feature.
Handles optional grouping & splitting.
}
