% Generated by roxygen2 (4.0.1): do not edit by hand
\name{D.ndim}
\alias{D.ndim}
\title{Calculate distance}
\usage{
D.ndim(point1, point2)
}
\arguments{
\item{point1}{The first input point}

\item{point2}{The secont input point.}
}
\description{
Calculate distance for any number of dimensions (without using a for loop)
Recall, Distance = sqrt((x2 - x1)^2 + (y2 - y1)^2 ... more terms here)
}

