% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.plot.R
\name{dif.plot}
\alias{dif.plot}
\title{Gera os graficos de DIF}
\usage{
dif.plot(EXP, SCO, main = NULL, groups = NULL, probs = c(0.05, 0.95),
  col.dif = NULL, density = "area", dir.create = "DIF",
  xlim = c(-4, 4), width = 9.6, height = 6.8, shinyDemo = NULL)
}
\arguments{
\item{EXP}{Objeto resultante da funcao calcDIF.}

\item{SCO}{Uma lista nomeada de acordo com os indices dos grupos em analise, contendo os percentis das distribuicoes de proficiencias.}

\item{main}{String ou vetor com o(s) titulo(s) dos graficos.}

\item{groups}{Vetor com o nome a ser plotado para cada um dos grupos, ordenado pelo indice do respectivo grupo.}

\item{probs}{Quantiles utilizados para o calculo dos intervalos de interesse.}

\item{col.dif}{Vetor com as cores a serem utilizadas nas curvas empiricas dos grupos.}

\item{density}{Plotar na imagem a densidade das proficiencias de cada grupo ('area', 'points'). Caso NULL, a densidade sera omitida.}

\item{dir.create}{Nome do diretorio a ser criado para guardar os graficos.}

\item{xlim}{Limites do eixo X a ser plotados.}

\item{width}{Largura em polegadas da imagem.}

\item{height}{Altura em polegadas da imagem.}

\item{shinyDemo}{Argumento auxiliar para o uso da funcao 'myR.app'.}
}
\description{
Cria um diretorio chamado 'DIF' e gera os graficos de Funcionamento Diferenciado Do Item (DIF).
}
\details{
Etapa Anterior: 'checkDIF'.
}
\examples{
EXP = remakeEXP(readLines("EXPtest.EXP"))

newGroup = c(6, 7)

EXP = calcDIF(EXP, newGroup)

SCO = read.SCO(readLines("SCOtest.SCO"))

SCO = lapply(split(SCO$SCO, SCO$Grupo), quantile, probs = c(.05, .95))

dif.plot(EXP, SCO)

}
\author{
Brunno Bittencourt
}
