% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Digit}
\alias{Digit}
\title{Digit}
\usage{
Digit(number, places)
}
\arguments{
\item{number}{a number to put into the function}

\item{places}{The number of places to make the number.}
}
\description{
Purpose of Digit is to take a number and place leading zeros before it if needed to bring the length of the number up to places.
Warning: Will not work properly with negative numbers
Inputs can be either numeric or character
Inputs must be integer values (e.g., "1" or 1 are okay, 1.5 is not.)
Places variable should be at least 1
Output is in character format (to preserve leading 0's!)
I have a similar function for Python, but it is implemented differently.
}

