% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitize_graph.R
\name{digitize_graph}
\alias{digitize_graph}
\title{Digitize graphs using package digitize}
\usage{
digitize_graph(fname, x_min, x_max, y_min, y_max, data_sets = 1,
  labels = 1:data_sets, log = c("", "x", "y", "xy"), plot = TRUE)
}
\arguments{
\item{fname}{Filename/ path of the graphic to read}

\item{x_min}{Minimum value on the horizontal axis}

\item{x_max}{Maximum value on the horizontal axis}

\item{y_min}{Minimum value on the vertical axis}

\item{y_max}{Maximum value on the vertical axis}

\item{data_sets}{defines how many data_sets will be extracted from the graph}

\item{labels}{character vector containing labels for data_sets}

\item{log}{defines logarithmic scales. By default scaling is assumed to be linear}

\item{plot}{If TRUE, the extracted data is plotted}
}
\value{
a data frame
}
\description{
Digitzes one or multiple datadata_sets within a single figure. One or both axes might be logarithmic. The user needs to define to values on both axes for calibration purposes. It´s not necessary to specify the global minimum or maximum values but coordinates that are possible to define precisely.
}
\references{
This functions is adopted from a blog post by Bart Rogiers:
http://rogiersbart.blogspot.de/2012/06/digitize-linear-and-semi-log-scale.html
}
\author{
Meinolf Ottensmann
}
