% Generated by roxygen2 (4.0.2): do not edit by hand
\name{distance}
\alias{distance}
\title{Compute distance}
\usage{
distance(x1, y1, x2, y2)
}
\arguments{
\item{x1}{the first x value}

\item{y1}{the first y value}

\item{x2}{the second x value}

\item{y2}{the second y value}
}
\value{
a scalar value
}
\description{
Compute distance
}
\details{
Sjekk funksjonen..
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}

