% Generated by roxygen2 (4.0.2): do not edit by hand
\name{distance3D}
\alias{distance3D}
\title{Compute distance 3D}
\usage{
distance3D(x1, y1, z1, x2, y2, z2)
}
\arguments{
\item{x1}{the first x value}

\item{y1}{the first y value}

\item{z1}{the first z value}

\item{x2}{the second x value}

\item{y2}{the second y value}

\item{z2}{the second z value}
}
\value{
a scalar value
}
\description{
Compute distance 3D
}
\details{
Sjekk funksjonen..
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}

