% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy.R
\name{doy}
\alias{doy}
\title{Day of year from date}
\usage{
doy(date = Sys.Date())
}
\arguments{
\item{date}{either object class date or a string of format "YYYY-MM-DD" that can be coerced into date format}
}
\description{
Day of year from date
}
\examples{
## day of current date
doy()

}
