% Generated by roxygen2 (4.0.2): do not edit by hand
\name{errormatrix}
\alias{errormatrix}
\title{Compute errormatrix}
\usage{
errormatrix(class, reference)
}
\arguments{
\item{class}{a classified vector}

\item{reference}{reference vector}
}
\value{
em errormatrix ,oa=oa,ua=ua,pa=pa,EM=EM,kappa=kappa)

oa overall accuracy

ua user's accuracy

pa producer's accuracy

kappa kappa-coefficient

EM errormatric printable
}
\description{
Algorithm to find peak values
}
\details{
Last update:29.10.2008  ??
}
\examples{
#Example: Congalton, R.G. (1991). A Review of Assessing the Accuracy of Classifications of Remotely Sensed Data. Remote Sensing of Environment, 37, 35-46:
x<-matrix(c(65,6,0,4,4,81,11,7,22,5,85,3,24,8,19,90),4,4)
xsup<-matrix(c(68,12,3,0,7,112,9,2,3,15,89,5,0,10,0,56),4,4)
xunsup<-matrix(c(60,15,6,2,11,102,13,4,3,14,90,5,4,8,2,52),4,4)
xmod<-matrix(c(75,4,3,1,6,116,7,1,1,11,96,4,0,3,2,61),4,4)
}
\author{
Hans Ole &Oslashrka \email{hans.ole.orka@gmail.com}
}
\references{
Congalton, R.G. (1991). A Review of Assessing the Accuracy of Classifications of Remotely Sensed Data. Remote Sensing of Environment, 37, 35-46

Story, M., & Congalton, R.G. (1986). Accuracy Assessment - a Users Perspective. Photogrammetric Engineering and Remote Sensing, 52, 397-399
}

