% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escore.R
\name{escore}
\alias{escore}
\title{Calcula uma medida de desempenho em um DF corrigido}
\usage{
escore(dadosCorr, method = "Escore")
}
\arguments{
\item{dadosCorr}{Banco de dados corrigido a ser utilizado. Este deve ser composto com os seguintes campos: (ID, CAD, PESO, IT1, ..., ITm).}

\item{method}{Metodo de resumo dos dados (Escore ou Normit).}
}
\value{
DF com a seguinte composicao: (ID, CAD, PESO, IT1, ..., ITn, ..., ITm, Desempenho).
}
\description{
Resume um banco de dados gerado pela funcao corrigir, atraves de uma medida de desempenho Escore/Normit.
}
\details{
Etapa Anterior: 'corrigir'.

Etapa Posterior: 'tct' , 'agi.plot'.
}
\examples{
bib = list(CAD1 = c(1,2), CAD2 = c(2,3), CAD3 = c(3,1))

nblocos = 3

b.len = 2

resp = data.frame(matrix(sample(c("A", "B", "C"), 40, replace = T), ncol = 4), stringsAsFactors = F); names(resp) = paste0("IT", 1:4)

dados = cbind(data.frame(ID = 1:10, CAD = paste0("CAD", sample(3, 10, replace = T)), PESO = 1), resp)

dados = vetor.extendido(dados, bib, nblocos, b.len)

dadosCorr = corrigir(dados, c("A", "B", "A", "C", "A", "B"))

dadosCorr = escore(dadosCorr)

}
\author{
Brunno Bittencourt
}
