% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filterBinomial}
\alias{filterBinomial}
\title{Creates a circular binomail filter}
\usage{
filterBinomial(w = 3, CircularRadius = "euclidan")
}
\arguments{
\item{w}{filter size}

\item{CircularRadius}{How to define the radius of the circular filter, either in a euclidian distance space (w-1)/2 or by using the pixels w/2. Defaults to "euclidian"}
}
\value{
a matrix of filter weights
}
\description{
Create a circular low-pass filter using binomial coefincient and filter weights obtained from  Pascal's triangle. Used for filtering Canopy height model before local maxima detection
}
\details{
The filter should be used toghter with the focal function in the raster-package.
}
\examples{
windows()
par(mfcol=c(2,2))
(f <- filterBinomial(w=3))
image(f)
(f <- filterBinomial(w=5))
image(f)
(f <- filterBinomial(w=9))
(f <- filterBinomial(w=19))
image(f)
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}
\seealso{
\link[raster]{focal} and \link{filterLocalMaxima}
}

