% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filterCircularOnes}
\alias{filterCircularOnes}
\title{Creates a circular filter}
\usage{
filterCircularOnes(w = 3, CircularRadius = "euclidan")
}
\arguments{
\item{w}{filter size}

\item{CircularRadius}{How to define the radius of the circular filter, either in a euclidian distance space (w-1)/2 or by using the pixels w/2. Defaults to "euclidian"}
}
\value{
a matrix zeros and ones
}
\description{
Create a circular filter of ones
}
\details{
The fitler should be used toghter with the focal function in the raster-package.
}
\examples{
windows()
(f <- filterCircularOnes(w=3))
image(f)
(f<- filterCircularOnes(w=5))
image(f)
(f <- filterCircularOnes(w=9))
(f <- filterCircularOnes(w=19))
image(f)
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}
\seealso{
\link[raster]{focal}
}

