% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filterLocalMaxima}
\alias{filterLocalMaxima}
\title{Local Maxima Filtering}
\usage{
filterLocalMaxima(CHM, w = 3, ov = 2)
}
\arguments{
\item{CHM}{a object of class SpatialGridDataFrame}

\item{w}{matrix of weights e.g. a 3 by 3 matrix; see Details. The matrix can also be expressed as the number of cells in a single direction or in two directions from the focal cell, in which case the weights are all set to 1. I.e. w=3 refers to a 3 by 3 matrix: 2 cells at each side of the focal cell, queen's case, 9 cells in total. This is equivalent to w=c(3,3). You can also specify a rectangular neighborhood, e.g. w=c(3,5); but the sides must be odd numbers. If you need even sides, you can add a column or row with weights of zero.}

\item{ov}{threshold of local maximas to consider}
}
\value{
object of class SpatialPointsDataFrame-class
}
\description{
Find local maxima's in a local neighborhod
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}

