% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{find_peaks}
\alias{find_peaks}
\title{Find local Maxima in a time series}
\usage{
find_peaks(df, y = "y")
}
\arguments{
\item{df}{a data frame}

\item{y}{variable name of the y coordinate}
}
\description{
Find local Maxima in a time series
}
