% Generated by roxygen2 (4.0.1): do not edit by hand
\name{force.hibernate}
\alias{force.hibernate}
\title{Force hibernate}
\usage{
force.hibernate()
}
\description{
Tell the computer to hibernate in a way that will still work if the machine is locked.
Based on code in the R package fun, except it hibernates and forces it.
}

