% Generated by roxygen2 (4.0.1): do not edit by hand
\name{force.shutdown}
\alias{force.shutdown}
\title{Force Shutdown}
\usage{
force.shutdown()
}
\description{
Tell the computer to shutdown, in a way that will still work if the machine is locked.
Based on code in the R package fun, except the wait option is replaced by the force option.
}

