% Generated by roxygen2 (4.0.2): do not edit by hand
\name{forwardSelectionJohn}
\alias{forwardSelectionJohn}
\title{Forward variable selection}
\usage{
forwardSelectionJohn(y, X, family = gaussian(), Crit = "p.value",
  pValueThreshold = NULL, pValueLikelihoodRatio = 0.05, trace = TRUE)
}
\arguments{
\item{y}{dependent variable}

\item{X}{a data.frame of x variables}

\item{family}{a description of the error distribution and link function to be used in the model. This can be a character string naming a family function, a family function or the result of a call to a family function. (See \code{family} for details of family functions.)}

\item{Crit}{Criterion to be used for variable selection "p.value" or "aic" implemented.}

\item{pValueThreshold}{a p-value threshold for single term models}

\item{pValueLikelihoodRatio}{a p-value threshold for the likelihood ratio test}

\item{trace}{logical printing model selection}
}
\value{
returns an object of class inheriting from "glm". See \code{glm}
}
\description{
Perform a forward variable selection procedure as described by John in a \code{glm} setting
}
\author{
Hans Ole &Oslashrka \email{hans.ole.orka@gmail.com}
}

