% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_list.R
\docType{data}
\name{genes_azi}
\alias{genes_azi}
\alias{genes_pbmc_new}
\alias{genes_skin}
\alias{geomx_rcc}
\alias{geomx_rcc2}
\alias{genes_gut}
\title{Azimuth (AZI) reference gene lists}
\format{
A named list of character vectors where each element is a vector of gene symbols.

An object of class \code{list} of length 10.

An object of class \code{list} of length 29.

An object of class \code{list} of length 21.

An object of class \code{list} of length 23.

An object of class \code{list} of length 38.
}
\source{
\url{https://azimuth.hubmapconsortium.org/references/human_pbmc/}
}
\usage{
genes_azi

genes_pbmc_new

genes_skin

geomx_rcc

geomx_rcc2

genes_gut
}
\description{
A named list containing gene sets used in Azimuth analysis.
https://azimuth.hubmapconsortium.org/references/human_pbmc/
}
\keyword{datasets}
