% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\docType{data}
\name{genes_azi}
\alias{genes_azi}
\title{Gene list for myR analyses
https://azimuth.hubmapconsortium.org/references/human_pbmc/}
\format{
A character vector of length 4.
}
\source{
Generated by package developer.
}
\usage{
genes_azi
}
\description{
A character vector of default gene symbols.
}
\keyword{datasets}
