% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\docType{data}
\name{genes_skin}
\alias{genes_skin}
\title{Gene list for myR analyses}
\format{
A character vector of length 4.
}
\source{
Generated by package developer.
}
\usage{
genes_skin
}
\description{
A character vector of default gene symbols.
}
\keyword{datasets}
