% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{get_default_plot_dims}
\alias{get_default_plot_dims}
\title{Get Default Plot Dimensions}
\usage{
get_default_plot_dims(plot_type = "default", n_items = NULL)
}
\arguments{
\item{plot_type}{Type of plot ("heatmap", "scatter", "boxplot", "default")}

\item{n_items}{Number of items (e.g., genes, clusters) - used for sizing}
}
\value{
List with width and height in inches
}
\description{
Returns sensible default plot dimensions based on plot type.
}
\examples{
get_default_plot_dims("heatmap", n_items = 20)

}
