% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{getmax}
\alias{getmax}
\title{Round up using a given digit type}
\usage{
getmax(x, digit)
}
\arguments{
\item{x}{input data}

\item{digit}{divisor number to be used}
}
\description{
Round up using a given digit type
}
