% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{getmin}
\alias{getmin}
\title{Round down using a given digit type}
\usage{
getmin(x, digit)
}
\arguments{
\item{x}{input data}

\item{digit}{divisor number to be used}
}
\description{
Round down using a given digit type
}
