% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getPolygonID}
\alias{getPolygonID}
\title{Polygons ids}
\usage{
getPolygonID(obj)
}
\arguments{
\item{obj}{a sp object}
}
\value{
returns an vector with IDs
}
\description{
Extraxt the ID of a SpatialPolygon or SpatialPolgyonsDataFrame. See \code{sp}
}
\author{
Hans Ole &Oslashrka \email{hans.ole.orka@gmail.com}
}

