% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gini}
\alias{gini}
\title{Gini coefficient}
\usage{
gini(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a scalar
}
\description{
Compute the Gini coefficient (Gini index or Gini ratio) which is a mesure of statistical dispersion
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}
\references{
Gini, C (1909) Concentration and dependency ratios (in Italian). English translation in Rivista di Politica Economica, 87 (1997), 769-789.
}

