% Generated by roxygen2 (4.0.1): do not edit by hand
\name{listtotext}
\alias{listtotext}
\title{list to text}
\usage{
listtotext(inlist, separator)
}
\arguments{
\item{inlist}{A list to be converted to text}

\item{separator}{The separator to separate the text}
}
\description{
Take a list (or vector) and convert it to text separated by separator.
This does not work for nested lists
}

