% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5ad_to_seurat.R
\name{load_h5ad_to_seurat_qs}
\alias{load_h5ad_to_seurat_qs}
\title{Load h5ad file and convert to Seurat object, then save as .qs file}
\usage{
load_h5ad_to_seurat_qs(
  h5ad_path,
  condaenv = "/home/jaecheon/miniconda3/envs/scvi-tools",
  save_path = NULL,
  save_name = NULL,
  use_layer = NULL,
  project_name = "h5ad_import"
)
}
\arguments{
\item{h5ad_path}{Path to the h5ad file}

\item{condaenv}{Conda environment with anndata installed (default: "/home/jaecheon/miniconda3/envs/scenvi")}

\item{save_path}{Directory to save the .qs file (default: current directory)}

\item{save_name}{Name for the saved file (without extension). If NULL, uses basename of h5ad_path}

\item{use_layer}{If h5ad has layers, which layer to use for counts (e.g., "counts", "RNA_counts"). If NULL, uses X matrix}

\item{project_name}{Project name for Seurat object (default: "h5ad_import")}
}
\description{
Load h5ad file and convert to Seurat object, then save as .qs file
}
