% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_utils.R
\name{load_intermediate}
\alias{load_intermediate}
\title{Load Intermediate Object}
\usage{
load_intermediate(filepath, log_list = NULL, input_dir = NULL)
}
\arguments{
\item{filepath}{Full path to input file}

\item{log_list}{Optional log list for logging}
}
\value{
Loaded object
}
\description{
Loads an R object saved with qs
}
