% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.label}
\alias{make.label}
\title{Make a numeric label}
\usage{
make.label(x, lbl, separ = "_")
}
\arguments{
\item{x}{The numeric part of the label}

\item{lbl}{The label part of the label}

\item{separ}{The separator for the label.}
}
\description{
make a numeric label in a way that can easily be used with apply
}

