% Generated by roxygen2 (4.0.1): do not edit by hand
\name{matrix.subset}
\alias{matrix.subset}
\title{Preserve matrix formatting when reformatting a matrix}
\usage{
matrix.subset(in.mat, index.var, target.value)
}
\arguments{
\item{in.mat}{The input matrix}

\item{index.var}{The variable to be used for subsetting}

\item{target.value}{The value of the subsetting variable to subset on.}
}
\description{
When a single column is extracted from a multicolumn matrix, R often strips the matrix
formatting. This generally causes my code to crash, because I assume that the formatting
is preserved. This function aims to preserve that formatting.
Testing note: Tested with an input matrix with 5 columns & 4 rows, and subset to obtain 0, 1 and 2 columns.
}

