% Generated by roxygen2 (4.0.2): do not edit by hand
\name{minError}
\alias{minError}
\title{Minimum error thresholding}
\usage{
minError(h)
}
\arguments{
\item{h}{histogram}
}
\value{
threshold
}
\description{
Find a global threshold for a grayscale image using the minimum error thresholding method.
}
\details{
Functions based on matlab function by 2004 Antti Niemist&#246 http://www.dentistry.bham.ac.uk/landinig/software/autothreshold/autothreshold.html
}
\examples{
x <- c(rnorm(100,10),rnorm(50,20))
h <- hist(x)
T <- minError(h)
hist(x)
abline(v=T)
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}
\references{
J. Kittler and J. Illingworth, "Minimum error thresholding," Pattern Recognition, vol. 19, pp. 41-47, 1986.

C. A. Glasbey, "An analysis of histogram-based thresholding algorithms,"CVGIP: Graphical Models and Image Processing, vol. 55, pp. 532-537, 1993.
}

