% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minor_ticks.R
\name{minor_ticks}
\alias{minor_ticks}
\title{insert minor axis ticks with blank label}
\source{
https://stackoverflow.com/questions/34533472/insert-blanks-into-a-vector-for-e-g-minor-tick-labels-in-r
}
\usage{
minor_ticks(x, nth, empty = TRUE, inverse = FALSE)
}
\arguments{
\item{x}{vector of main axis labels}

\item{nth}{number of blanks between two elements of major_labs}

\item{empty}{logigcal}

\item{inverse}{logical}
}
\description{
insert minor axis ticks with blank label
}
