% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Mode: Calculates the mode of a vector}
\usage{
Mode(x, na.rm = TRUE)
}
\arguments{
\item{x}{Vector of numbers or characters}

\item{na.rm}{Boolean indicating to remove NA values}
}
\description{
Mode: Calculates the mode of a vector
}
\details{
Ignores NA, if na.rm = TRUE
}
