% Generated by roxygen2 (4.0.2): do not edit by hand
\name{closing}
\alias{closing}
\alias{convert2BinaryRaster}
\alias{dilation}
\alias{erosion}
\alias{opening}
\title{Moprpholgogical operators : closing}
\usage{
closing(r)

convert2BinaryRaster(r)

dilation(r)

erosion(r)

opening(r)
}
\arguments{
\item{r}{a raster object}
}
\value{
returns an object of class raster. See \code{raster}
}
\description{
Perform a closing

Convert a raster to a binary raster

Perform a dilation

Perform a erosion

Perform a opening
}
\author{
Hans Ole &Oslashrka \email{hans.ole.orka@gmail.com}
}
\seealso{
\link[raster]{focal}
}

