% Generated by roxygen2 (4.0.1): do not edit by hand
\name{move.files}
\alias{move.files}
\title{Moves files for building a package}
\usage{
move.files(my.pkg)
}
\arguments{
\item{my.pkg}{The package of interest. Currently this only does something for
the spatialdemography package!}
}
\description{
There are a number of files in my base R directory that I don't want in the
R package build. .Rbuildignore does not seem to work for me.
So, I wrote this function and also return.files to expedite things
}

