% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{mybox_cell}
\alias{mybox_cell}
\title{세포 단위 피처 발현 박스플롯 (plot_gene_boxplot 스타일)}
\usage{
mybox_cell(
  data,
  features,
  group.by,
  split.by,
  idents = NULL,
  assay = NULL,
  layer = "data",
  ncol = 3,
  pt.size = 0.2,
  violin = FALSE,
  add_stats = TRUE,
  ...
)
}
\arguments{
\item{data}{Seurat 객체 또는 data.frame}

\item{features}{플로팅할 피처 (유전자 또는 메타데이터)}

\item{group.by}{패널(facet)로 분리할 그룹 변수 (예: "cell_type")}

\item{split.by}{X축에서 나눌 그룹 변수 (예: "timepoint")}

\item{idents}{'group.by'에서 포함할 특정 값들}

\item{assay}{Seurat 객체용 assay}

\item{layer}{Seurat 객체용 layer}

\item{ncol}{플롯 컬럼 수}

\item{pt.size}{점 크기 (0이면 숨김)}

\item{violin}{바이올린 플롯 오버레이}

\item{add_stats}{통계 검정 추가 (ggpubr::stat_compare_means)}

\item{...}{stat_compare_means에 전달할 추가 인수 (예: paired = TRUE, method = "t.test")}
}
\value{
ggplot 객체
}
\description{
세포 단위 피처 발현 박스플롯 (plot_gene_boxplot 스타일)
}
