% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{myhm_genes2}
\alias{myhm_genes2}
\title{title = "Normalized Gene Set Expression per Cluster",
x = "Cluster",
y = "Gene Set"}
\usage{
myhm_genes2(
  sobj,
  group = "seurat_clusters",
  value = "average",
  assay = "SCT",
  genes = NULL,
  title = "Normalized Gene Expression per Cluster",
  x_label = "Cluster",
  y_label = "Genes"
)
}
\description{
title = "Normalized Gene Set Expression per Cluster",
x = "Cluster",
y = "Gene Set"
}
