% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{myhm_genesets2}
\alias{myhm_genesets2}
\title{awesome function for making heatmap(tile plot) of gene modules}
\usage{
myhm_genesets2(
  sobj,
  group = "seurat_clusters",
  value = "average",
  assay = "SCT",
  gene_sets = NULL,
  title = "Normalized Gene Set Expression per Cluster",
  x_label = "Cluster",
  y_label = "Gene Set"
)
}
\description{
awesome function for making heatmap(tile plot) of gene modules
}
