% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{myline}
\alias{myline}
\title{단일 벡터 line plot with cutoff intersection}
\usage{
myline(
  vector,
  main = "Sorted Values",
  ylab = "Value",
  cutoff = NULL,
  cutoff_col = "red",
  show_intersection = TRUE,
  ...
)
}
\description{
단일 벡터 line plot with cutoff intersection
}
