% Generated by roxygen2 (4.0.2): do not edit by hand
\name{myLMfitfunction}
\alias{myLMfitfunction}
\title{BEST SUBSET VARIABLE SELECTION WITH PENALIZATION FOR COLLINEARITY}
\usage{
myLMfitfunction(Xy, Nvmax = 5)
}
\arguments{
\item{Xy}{data.frame with X variables and the y variable named Xy$y}

\item{Nvmax}{maximum size of subsets to examine}
}
\value{
a object of type lm
}
\description{
Model selection using the Bayesian information criterion (BIC) and a branc and bound
search for the best subset implemented in package leaps.Among the alternative subsets, the subset with lowest BIC are preferred.
Furthermore, the selection procedure penalized for collinearity using the
variance inflation factor (VIF). The subset with one variable less are iteratively
selected if any of the variables in the current subset has a large VIF greater than 5.
TODO: implement posibility to change VIF value
}
\details{
Check function...
}
\author{
Hans Ole {\O}rka \email{hans.ole.orka@gmail.org}
}

