% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myR.app.R
\name{myR.app}
\alias{myR.app}
\title{Interface Point-Click do myR}
\usage{
myR.app()
}
\description{
Abre uma interface para se utilizar as funcoes disponiveis no pacote.
}
\examples{
myR.app()

}
\author{
Brunno Bittencourt
}
