% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myr.library}
\alias{myr.library}
\title{First line is title}
\usage{
myr.library(package)
}
\description{
This is a brief description.
}
\details{
Anything else after the description goes to the Details section.
}
