% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{myr_library}
\alias{myr_library}
\title{First line is title}
\usage{
myr_library(package)
}
\arguments{
\item{x}{a numeric vector giving the values of the random variable.}

\item{probs}{optional, a numeric vector giving the probabilities
corresponding to each x value. If not specific, assumes all outcomes
are equally likely}
}
\value{
An S3 object of class rv.
}
\description{
This is a brief description.
}
\details{
Anything else after the description goes to the Details section.
}
