% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{myr_package}
\alias{myr_package}
\alias{myr_package-package}
\title{myr (package)}
\description{
Package containing assorted useful functions
}
\details{
\tabular{ll}{ %% Note tabular{ll} is tabular{lowercase(LL)} not tabular{11}
Package: \tab myr\cr
Type: \tab Package\cr
Version: \tab 0.1.6\cr
Date: \tab 2015-09-16\cr
License: \tab GPL-2 (or later)\cr
}
}
\author{
Alexander "Sasha" Keyel
Maintainer: Sasha Keyel <skeyel@gmail.com>
}
\keyword{package}

