% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openstring.R
\name{openstring}
\alias{openstring}
\title{Abre strings de respostas}
\usage{
openstring(dados)
}
\arguments{
\item{dados}{Vetor contendo as strings de respostas.}
}
\value{
DF com a seguinte composicao: (IT1, ..., ITn, ..., ITm).
}
\description{
Gera uma data.frame com as respostas provindas de um vetor de strings.
}
\examples{
resp = c("AABDAB", "DBDADD", "CCBDAD")

resp = openstring(resp)

}
\author{
Brunno Bittencourt
}
