% Generated by roxygen2 (4.0.2): do not edit by hand
\name{otsu}
\alias{otsu}
\title{Otsu thresholding version 2}
\usage{
otsu(y, m)
}
\arguments{
\item{y}{vector of values}

\item{m}{number of modes - 1 (two modes, m=1)}
}
\value{
threshold
}
\description{
Find otsu threshold
}
\details{
to be written an other time...
}
\author{
Hans Ole &Oslashrka \email{hans.ole.orka@gmail.org}
}

