% Generated by roxygen2 (4.0.2): do not edit by hand
\name{otsuModif}
\alias{otsuModif}
\title{Otsu thresholding}
\usage{
otsuModif(h)
}
\arguments{
\item{h}{histogram created by \code{hist} or \code{sshist}}
}
\value{
threshold
}
\description{
Find otsu threshold
}
\details{
to be written an other time...
}
\examples{
x <- c(rnorm(100,10),rnorm(50,20))
h <- hist(x)
T <- otsuModif(h)
hist(x)
abline(v=T)
}
\author{
Hans Ole &Oslashrka \email{hans.ole.orka@gmail.org}
}

