% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_deg_cursor.R
\name{pb_utils_check_sample_distribution}
\alias{pb_utils_check_sample_distribution}
\title{Check sample distribution per group}
\usage{
pb_utils_check_sample_distribution(
  meta,
  grouping_col,
  patient_col,
  min_reps = 2,
  verbose = TRUE
)
}
\description{
Check sample distribution per group
}
\keyword{internal}
