% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_deg_cursor.R
\name{pb_utils_create_design}
\alias{pb_utils_create_design}
\title{Create design matrix from formula and metadata}
\usage{
pb_utils_create_design(formula_str, data, group_col = NULL)
}
\description{
Create design matrix from formula and metadata
}
\keyword{internal}
