% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pb_deg_cursor.R
\name{pb_utils_map_sample_group}
\alias{pb_utils_map_sample_group}
\title{Create sample-group mapping from Seurat metadata}
\usage{
pb_utils_map_sample_group(seurat_obj, sample_col, group_col, verbose = TRUE)
}
\description{
Create sample-group mapping from Seurat metadata
}
\keyword{internal}
