% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pcv}
\alias{pcv}
\title{Point Cloud Viewer - Edited version}
\usage{
pcv(pc, size = 1, col = "default", add = FALSE, background = "black",
  col.column = 3, type = "p", alpha = 1, device = rgl.cur(),
  grey.out = NA, windowRect = c(40, 70, 1000, 1000), ...)
}
\arguments{
\item{pc}{a matrix with xyz s in coloumns 1 2 3}

\item{size}{something other than something else}

\item{col}{a iteger with the index to the coloumn  or the name of the column used for coloring points}

\item{add}{logical it adding the}

\item{background}{Color of background}

\item{col.column}{number or name of column used to color points}

\item{type}{kind of type}

\item{alpha}{something}

\item{device}{exisitng device ?}

\item{grey.out}{something else}

\item{windowRect}{someting more}

\item{...}{other issues to be pased}
}
\description{
Visualize point colud using rgl
}
\author{
Marius and Hans Ole
}

