% Generated by roxygen2 (4.0.2): do not edit by hand
\name{peaks}
\alias{peaks}
\title{Peak values}
\usage{
peaks(series, span = 3, do.pad = TRUE)
}
\arguments{
\item{series}{the values to find peaks in}

\item{span}{a span parameter}

\item{do.pad}{= TRUE}
}
\value{
a logical vector idicating if the point is a peak value or not
}
\description{
Algorithm to find peak values
}
\examples{
set.seed(7)
y <- rpois(100, lambda = 7)
py <- peaks(y)
plot(y, type="o", cex = 1/4, main = "y and peaks(y,3)")
points(seq(y)[py], y[py], col = 2, cex = 1.5)

p7 <- peaks(y,7)
points(seq(y)[p7], y[p7], col = 3, cex = 2)
mtext("peaks(y,7)", col=3)

set.seed(2)
x <- round(rnorm(500), 2)
y <- cumsum(x)

plot(y, type="o", cex = 1/4)
p15 <- peaks(y,15)
points(seq(y)[p15], y[p15], col = 3, cex = 2)
mtext("peaks(y,15)", col=3)
}
\author{
R community  ???
}
\references{
https://stat.ethz.ch/pipermail/r-help/2005-November/083376.html
}

