% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_frequency.R
\name{perform_prior_tests}
\alias{perform_prior_tests}
\title{Perform prior tests for statistical assumptions}
\usage{
perform_prior_tests(data, value_col, group_col)
}
\arguments{
\item{data}{Data frame with sample-level data}

\item{value_col}{Column name for the values to test}

\item{group_col}{Column name for grouping variable}
}
\value{
List with test results
}
\description{
Perform prior tests for statistical assumptions
}
