% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tml.R
\name{plot_tml_metrics}
\alias{plot_tml_metrics}
\title{Plot TML Model Comparison Metrics}
\usage{
plot_tml_metrics(tml_object, metric = NULL)
}
\arguments{
\item{tml_object}{A result object from \code{\link[=TML7]{TML7()}}.}

\item{metric}{The metric to plot (e.g., "ROC", "Sens", "Spec"). Default is the best metric used in training.}
}
\value{
A ggplot object.
}
\description{
Visualizes the performance metrics of different L2 models across cross-validation folds.
}
