% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_working.R
\name{PlotClusterProportions}
\alias{PlotClusterProportions}
\title{클러스터 비율 비교 박스플롯 생성 (Pseudobulk 방식)}
\usage{
PlotClusterProportions(
  sobj,
  patient_col,
  group.by,
  split.by,
  test.use = "wilcox.test",
  p.adj = "BH",
  split.na.use = FALSE,
  colors = NULL
)
}
\arguments{
\item{sobj}{Seurat 객체}

\item{patient_col}{메타데이터 내 환자 또는 샘플 ID 컬럼명}

\item{group.by}{비교할 클러스터 컬럼명}

\item{split.by}{비교할 그룹 컬럼명}

\item{test.use}{사용할 통계 검정 ("wilcox.test", "t.test", "kruskal.test", "anova")}

\item{p.adj}{p-value 조정 방법 ("BH", "bonferroni", "none" 등)}

\item{split.na.use}{'split.by' 컬럼의 NA 값 처리 방법 (FALSE: 제외, TRUE: 별도 그룹으로 취급)}

\item{colors}{그룹별로 사용할 색상 팔레트 (선택 사항)}
}
\value{
ggplot 객체와 통계 결과가 포함된 리스트
}
\description{
클러스터 비율 비교 박스플롯 생성 (Pseudobulk 방식)
}
