% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.model.R
\name{pos.model}
\alias{pos.model}
\title{Calcula o posicionamento via modelo}
\usage{
pos.model(pars, breaks = c(-5, 5, 0.1), crit = list(PROP = 0.65, PROP2
  = 1, DIF = 0))
}
\arguments{
\item{pars}{DF com os parametros dos itens. Deve ser composto com os seguintes campos: (a, b, c).}

\item{breaks}{Limite inferior e superior dos pontos de posicionamento, acompanhados pelo step da sequencia.}

\item{crit}{Lista com os criterios para posicionamento.
\itemize{
 \item{PROP}{: Proporcao minima de respostas corretas no nivel}.
 \item{PROP2}{: Proporcao maxima de respostas corretas no nivel anterior}.
 \item{DIF}{: Diferenca minima entre PROP e PROP2}.
}}
}
\value{
Vetor com o posicionamento (se possivel) de cada item.
}
\description{
Calcula o posicionamento dos itens atraves do modelo 3PL.
}
\examples{
bib = list(CAD1 = c(1,2), CAD2 = c(2,3), CAD3 = c(3,1))

nblocos = 3

b.len = 2

resp = data.frame(matrix(sample(c("A", "B", "C"), 40, replace = T), ncol = 4), stringsAsFactors = F); names(resp) = paste0("IT", 1:4)

dados = cbind(data.frame(ID = 1:10, CAD = paste0("CAD", sample(3, 10, replace = T)), PESO = 1), resp)

dados = vetor.extendido(dados, bib, nblocos, b.len)

dadosCorr = corrigir(dados, c("A", "B", "A", "C", "A", "B"))

dadosCorr = escore(dadosCorr)

pars = data.frame(Codigo = 1001:1006, a = abs(rnorm(6, mean = 1.5, sd = 0.5)), b = rnorm(6, mean = 0.5), c = runif(6, 0.1, 0.3))
			  
POS = pos.model(pars, crit = list(PROP = 0.6, DIF = 0.05))

}
\seealso{
pos.prop()
}
\author{
Brunno Bittencourt
}
