% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_utils.R
\name{print.gene_signature}
\alias{print.gene_signature}
\title{Print Method for Gene Signature Objects}
\usage{
\method{print}{gene_signature}(x, n, ...)
}
\arguments{
\item{x}{A gene_signature object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Gene Signature Objects
}
