% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_working.R
\name{PrintTopMarkersForLLM}
\alias{PrintTopMarkersForLLM}
\title{FindAllMarkers 결과에서 Top N 유전자를 LLM용으로 출력}
\usage{
PrintTopMarkersForLLM(
  markers_df,
  top_n = 50,
  cluster_col = "cluster",
  gene_col = "gene",
  p_val_col = "p_val",
  p_val_adj_col = "p_val_adj",
  avg_log2FC_col = "avg_log2FC"
)
}
\arguments{
\item{markers_df}{FindAllMarkers()의 결과인 데이터프레임}

\item{top_n}{각 클러스터별로 추출할 상위 유전자 개수}

\item{cluster_col}{'cluster' 정보가 있는 컬럼명 (기본값: "cluster")}

\item{gene_col}{'gene' 정보가 있는 컬럼명 (기본값: "gene")}

\item{p_val_col}{p-value 컬럼명 (기본값: "p_val")}

\item{p_val_adj_col}{p-value (adjusted) 컬럼명 (기본값: "p_val_adj")}

\item{avg_log2FC_col}{log-fold change 컬럼명 (기본값: "avg_log2FC")}
}
\value{
콘솔에 클러스터별 Top N 유전자 목록을 출력합니다.
}
\description{
FindAllMarkers 결과에서 Top N 유전자를 LLM용으로 출력
}
