% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_demulti.R
\name{process_demuxalot}
\alias{process_demuxalot}
\title{Process Demuxalot Output (SNP-based)}
\usage{
process_demuxalot(
  demux_file,
  sample_names,
  barcode_col = "BARCODE",
  singlet_threshold = 0.5,
  doublet_threshold = 0.3,
  gem_name,
  gem_index = 1
)
}
\arguments{
\item{demux_file}{Path to demuxalot posterior CSV file}

\item{sample_names}{Vector of sample names (will be used as column names)}

\item{barcode_col}{Column name for barcodes (default: "BARCODE")}

\item{singlet_threshold}{Minimum probability for singlet (default: 0.5)}

\item{doublet_threshold}{Minimum probability for doublet (default: 0.3)}

\item{gem_name}{GEM name to add to metadata}

\item{gem_index}{Index for creating unique barcodes (default: 1)}
}
\value{
Data frame with barcode mapping including:
\item{Barcode}{Unique barcode ID}
\item{Best_Sample}{Assigned sample(s)}
\item{droplet_demulti}{singlet_demulti or doublet_demulti}
\item{GEM}{GEM name}
}
\description{
Processes demuxalot posterior CSV file and creates barcode mapping
}
