% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_demulti.R
\name{process_hto_demux}
\alias{process_hto_demux}
\title{Process HTO Demultiplexing}
\usage{
process_hto_demux(
  obj,
  method = "HTODemux",
  assay_name = "Multiplexing Capture",
  positive_quantile = 0.99,
  sample_name,
  gem_name
)
}
\arguments{
\item{obj}{Seurat object with HTO assay}

\item{method}{Demultiplexing method: "HTODemux" or "MULTIseqDemux"}

\item{assay_name}{Name of HTO assay (default: "Multiplexing Capture")}

\item{positive_quantile}{Quantile for positive HTO signal (default: 0.99)}

\item{sample_name}{Sample name to add to metadata}

\item{gem_name}{GEM name to add to metadata}
}
\value{
Seurat object with demultiplexing results in metadata
}
\description{
Performs HTO demultiplexing using Seurat's HTODemux or MULTIseqDemux
}
