% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_lab.R
\name{pval_lab}
\alias{pval_lab}
\title{Add p.value to plot with sensible precision}
\usage{
pval_lab(p.value, digits = 2, significance = F)
}
\arguments{
\item{p.value}{numeric value}

\item{digits}{integer}

\item{significance}{logical}
}
\description{
Add p.value to plot with sensible precision
}
