% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.excel}
\alias{read.excel}
\title{Read workbook from excel files}
\usage{
read.excel(filename, header = TRUE, sheet = 1)
}
\arguments{
\item{filename}{full path to excel file.}

\item{header}{logical value for including the header from the excel sheet or not}

\item{sheet}{a numeric value indicating which Sheet to be read}
}
\value{
data.frame
}
\description{
Many people use excel files to store data in. This function reads a specified excel worksheet into a data.frame.
}
\author{
Marius Hauglin & Hans Ole &Oslashrka
}
\seealso{
\code{\link{XLConnect}}
}

