% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.SCO.R
\name{read.SCO}
\alias{read.SCO}
\title{Le um arquivo .SCO}
\usage{
read.SCO(SCO, summary = TRUE, getID = FALSE)
}
\arguments{
\item{SCO}{Nome do arquivo a ser lido.}

\item{summary}{Printar 'table' dos grupos e 'summary' das proficiencias para fins de consistencia?}

\item{getID}{Manter IDs no resultado.}
}
\value{
DF com o grupo e a proficiencia de cada individuo do arquivo.
}
\description{
Le e manipula um arquivo .SCO para a forma requerida na etapa posterior.
}
\details{
Etapa Posterior: 'cci.plot'.
}
\examples{
SCO = read.SCO(readLines("SCOtest.SCO"))

}
\author{
Brunno Bittencourt
}
