% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remakeEXP.R
\name{remakeEXP}
\alias{remakeEXP}
\title{Le um arquivo .EXP}
\usage{
remakeEXP(EXP)
}
\arguments{
\item{EXP}{Objeto resultante da funcao readLines {base}.}
}
\value{
Lista com um DF reestruturado para cada item do arquivo.
}
\description{
Le e manipula um arquivo .EXP para a forma requerida na etapa posterior.
}
\details{
Etapa Posterior: 'calcDIF'.
}
\examples{
EXP = remakeEXP(readLines("EXPtest.EXP"))

}
\author{
Brunno Bittencourt
}
