% Generated by roxygen2 (4.0.1): do not edit by hand
\name{return.files}
\alias{return.files}
\title{Return files to package directory}
\usage{
return.files(my.pkg)
}
\arguments{
\item{my.pkg}{The package of interest, currently this is only scripted for
spatialdemography!}
}
\description{
This returns the files that were removed with the move.files function.
This could probably be turned into a single function, where the user
just specifies the direction of transfer!
}

