% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revcomp_seq.R
\name{revcomp_seq}
\alias{revcomp_seq}
\title{Return the reverse complement of a DNA sequence}
\usage{
revcomp_seq(x)
}
\arguments{
\item{x}{a DNA sequence}
}
\description{
Return the reverse complement of a DNA sequence
}
