% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error}
\usage{
rmse(y, yhat)
}
\arguments{
\item{y}{the reference y value}

\item{yhat}{the estimated y valye}
}
\value{
a scalar value
}
\description{
Compute the root means square error
}
\author{
Hans Ole Orka \email{hans.ole.orka@gmail.org}
}

