% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudobulk_deg.R
\name{run_DESeq2_pseudobulk_advanced}
\alias{run_DESeq2_pseudobulk_advanced}
\title{Run advanced DESeq2 analysis on pseudobulk data}
\usage{
run_DESeq2_pseudobulk_advanced(
  pb_matrix,
  metadata,
  group.by,
  ident.1,
  ident.2 = NULL,
  covariates = NULL,
  design.formula = NULL,
  logfc.threshold = 0.25
)

run_DESeq2_pseudobulk_advanced(
  pb_matrix,
  metadata,
  group.by,
  ident.1,
  ident.2 = NULL,
  covariates = NULL,
  design.formula = NULL,
  logfc.threshold = 0.25
)
}
\description{
Run advanced DESeq2 analysis on pseudobulk data

Run advanced DESeq2 analysis on pseudobulk data
}
