% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{runMAST}
\alias{runMAST}
\title{MAST 파이프라인 함수}
\usage{
runMAST(sobj, formula, min_cells_expr = 10, n_cores = 4, lrt_variable = NULL)
}
\arguments{
\item{sobj}{(Seurat) Seurat 객체}

\item{formula}{(formula or character) lme4 문법의 포뮬러}

\item{min_cells_expr}{(numeric) 유전자 필터링 기준 (최소 발현 세포 수)}

\item{n_cores}{(numeric) 병렬 처리 코어 수}

\item{lrt_variable}{(character) LRT 검정을 수행할 변수명 (예: "type")}
}
\description{
\link{수정} Seurat -> SCE로 변환하여 MAST를 실행
}
