% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cursor.R
\name{runMAST_v1}
\alias{runMAST_v1}
\title{MAST 파이프라인 함수}
\usage{
runMAST_v1(
  sobj,
  formula,
  min_cells_expr = 10,
  n_cores = 4,
  lrt_variable = NULL
)
}
\arguments{
\item{sobj}{(Seurat) Seurat 객체}

\item{formula}{(formula) lme4 문법의 포뮬러.
             (예: ~ g3 + (1|hos_no) + (1|GEM))}

\item{min_cells_expr}{(numeric) 유전자 필터링 기준 (최소 발현 세포 수)}

\item{n_cores}{(numeric) 병렬 처리 코어 수}

\item{lrt_variable}{(character) LRT 검정을 수행할 변수명 (예: "g3")}
}
\value{
(data.table) MAST 요약 결과
}
\description{
lme4의 다중 임의 효과를 지원하는 zlm (Hurdle-LMM) 실행
}
