% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cursor.R
\name{runMUSCAT_cursor_legacy}
\alias{runMUSCAT_cursor_legacy}
\title{Muscat 파이프라인 함수 v1 (DEPRECATED)}
\usage{
runMUSCAT_cursor_legacy(...)
}
\description{
이 함수는 deprecated되었습니다. test_analysis.R의 runMUSCAT_v5 또는 runMUSCAT2_v1을 사용하세요.
}
