% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cursor.R
\name{runMUSCAT_v1}
\alias{runMUSCAT_v1}
\title{Muscat (Pseudo-bulking) 파이프라인 함수}
\usage{
runMUSCAT_v1(
  sobj,
  cluster_id = "cell_type",
  sample_id = "hos_no",
  group_id = "g3",
  formula_str = "~ 0 + group + batch",
  contrast = NULL,
  method = "dream"
)
}
\arguments{
\item{sobj}{(Seurat) Seurat 객체. meta.data에 아래 3개 컬럼 포함 필수}

\item{cluster_id}{(character) 세포 유형 컬럼명 (예: "cell_type")}

\item{sample_id}{(character) 환자/샘플 ID 컬럼명 (예: "hos_no")}

\item{group_id}{(character) 비교할 그룹 컬럼명 (예: "g3")}

\item{formula_str}{(character) DE 분석에 사용할 포뮬러 (limma/edgeR/DESeq2)
                (예: "~ 0 + group_id + batch_col")}

\item{contrast}{(character) 비교할 contrast (limma::makeContrasts 문법)
                (예: "group_idLevelA-group_idLevelB")}

\item{method}{(character) DE 엔진: "dream", "edgeR", "DESeq2"}
}
\value{
(list) muscat::pbDS 결과 (세포 유형별 DE 테이블 리스트)
}
\description{
세포 유형별, 환자별로 count를 합산(Pseudo-bulk)하여 DE 분석
}
