% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cursor.R, R/test_working.R
\name{runNEBULA}
\alias{runNEBULA}
\title{NEBULA 파이프라인 함수}
\usage{
runNEBULA(
  sobj,
  layer = "counts",
  fixed_effects = c("target_col", "celltype_col"),
  covar_effects = c("batch_col"),
  patient_col = "patient_col",
  offset = "nCount_RNA",
  min_count = 10
)

runNEBULA(
  sobj,
  layer = "counts",
  fixed_effects = c("target_col", "celltype_col"),
  covar_effects = c("batch_col"),
  patient_col = "patient_col",
  offset = "nCount_RNA",
  min_count = 10
)
}
\arguments{
\item{sobj}{(Seurat) Seurat 객체}

\item{layer}{(character) Raw count가 저장된 layer (예: "counts")}

\item{fixed_effects}{(character vector) 고정 효과로 사용할 meta.data 컬럼명
                   (예: c("target_col", "celltype_col"))}

\item{covar_effects}{(character vector) 임의 효과처럼 보정하고 싶은 공변량.
                   NEBULA의 한계로 인해 '고정 효과'로 처리됩니다.
                   (예: c("batch_col", "percent.mt"))}

\item{patient_col}{(character) 환자 ID 컬럼. NEBULA의 'id' (유일한 임의 효과)로 사용됨.}

\item{offset}{(character) Offset으로 사용할 'numeric' 컬럼 (예: "nCount_RNA")}

\item{min_count}{(numeric) 유전자 필터링 기준 (최소 발현 세포 수)}
}
\value{
(nebula) NEBULA 실행 결과 객체

(nebula) NEBULA 실행 결과 객체
}
\description{
Seurat 객체를 받아 NA 처리, 데이터 정렬, NEBULA GLMM을 실행.
Modeling에서, y~NB(mu,phi)이며, log(mu)=log(o)+BX+random effect 에서 "o"는 offset, random effect는 patient_col, X는 fixed effects와 covar_effects.

Seurat 객체를 받아 NA 처리, 데이터 정렬, NEBULA GLMM을 실행.
Modeling에서, y~NB(mu,phi)이며, log(mu)=log(o)+BX+random effect 에서 "o"는 offset, random effect는 patient_col, X는 fixed effects와 covar_effects.
}
