% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cursor.R
\name{runNEBULA_cursor_legacy}
\alias{runNEBULA_cursor_legacy}
\title{NEBULA 파이프라인 함수 (DEPRECATED)}
\usage{
runNEBULA_cursor_legacy(...)
}
\description{
이 함수는 deprecated되었습니다. test_analysis.R의 runNEBULA_v1을 사용하세요.
}
