% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sample.fixed}
\alias{sample.fixed}
\title{Sample (fixed)}
\usage{
sample.fixed(x, size)
}
\arguments{
\item{x}{The vector to be sampled}

\item{size}{The number of samples to take}
}
\description{
Default behavior of this function was unacceptable, so I fixed it
if length(x) == 1, returns x instead of using x as a number to be sampled from.
}

