% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{sanitize_filename}
\alias{sanitize_filename}
\title{Sanitize Filename}
\usage{
sanitize_filename(filename, replacement = "_")
}
\arguments{
\item{filename}{Filename to sanitize}

\item{replacement}{Character to replace problematic characters with (default: "_")}
}
\value{
Sanitized filename
}
\description{
Removes or replaces characters that are problematic in filenames.
}
\examples{
sanitize_filename("my/file:name?.txt")
# Returns: "my_file_name_.txt"

}
