% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_utils.R
\name{save_intermediate}
\alias{save_intermediate}
\title{Save Intermediate Object}
\usage{
save_intermediate(obj, filepath, log_list = NULL, prevent_overwrite = TRUE)
}
\arguments{
\item{obj}{Object to save}

\item{filepath}{Full path to output file}

\item{log_list}{Optional log list for logging}

\item{prevent_overwrite}{If TRUE, appends _1, _2, ... if file exists (default: TRUE)}
}
\value{
Actual filepath used (may differ if prevent_overwrite is TRUE)
}
\description{
Saves an R object using qs for fast serialization
Automatically appends _1, _2, ... if file already exists to prevent overwriting
}
