% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{scatter_smooth}
\alias{scatter_smooth}
\title{Statistical Analysis for scRNA-seq Metadata}
\usage{
scatter_smooth(
  sobj,
  feature,
  clinical_variable = "nih_change",
  transpose = FALSE
)
}
\arguments{
\item{transpose}{whether to... transpose}
}
\value{
a simple form of scatter plot with regression fit.
}
\description{
This function performs comprehensive statistical analysis on scRNA-seq data
comparing groups across different categorical variables.
}
