% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_frequency.R
\name{seurat_posthoc_analysis}
\alias{seurat_posthoc_analysis}
\title{Post-hoc Analysis for scRNA-seq Statistical Tests}
\usage{
seurat_posthoc_analysis(data, results, alpha = 0.05)
}
\arguments{
\item{data}{A Seurat object or metadata dataframe}

\item{results}{Results dataframe from seurat_group_stats function}

\item{alpha}{Significance level (default: 0.05)}
}
\value{
A list of post-hoc test results
}
\description{
This function performs post-hoc analysis based on results from seurat_group_stats.
}
