% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sshist}
\alias{sshist}
\title{Optimal number of bins of a histogram}
\usage{
sshist(x)
}
\arguments{
\item{x}{Sample data.
x_min The data range within which the optimization applies to. Default value is min(x).
 x_max The data range within which the optimization applies to. Default value is max(x).
A vector that specifies number of divisions (bins). The optimal number of bins is selected from the elements of N.  Default value is N = 2:50.}
}
\value{
Output argument
}
\description{
Function `sshist' returns the optimal number of bins of a histogram. The optimization principle is L2 loss function. See ref. for details.
}
\author{
Author Hideaki Shimazaki \email{shimazaki at jhu.edu}
2006 Author Hideaki Shimazaki
}
\references{
Shimazaki and Shinomoto, Neural Comput. 19(6), 1503-1527, 2007
}

