% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ste}
\alias{ste}
\title{Calculate standard error}
\usage{
ste(values)
}
\arguments{
\item{values}{the values to calcualte the standard error for}
}
\description{
I thought this was supposed to be divided by sqrt of n-1, but wikipedia, Stackoverflow, and Zar p 72 all say sqrt n.
}

