% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_demulti.R
\name{summarize_demulti_results}
\alias{summarize_demulti_results}
\title{Summarize Demultiplexing Results}
\usage{
summarize_demulti_results(assignments)
}
\arguments{
\item{assignments}{Data frame from get_barcode_mapping}
}
\value{
List containing:
\item{n_total}{Total number of cells}
\item{n_singlets}{Number of singlet assignments}
\item{n_doublets}{Number of doublet assignments}
\item{n_negative}{Number of negative/unassigned cells}
\item{doublet_rate}{Proportion of doublets}
\item{assignment_rate}{Proportion successfully assigned}
}
\description{
Generates summary statistics for demultiplexing results.
}
