% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_claude.R
\name{summarize_geomx_object}
\alias{summarize_geomx_object}
\title{Quick Summary of Seurat Object for GeoMx Data}
\usage{
summarize_geomx_object(seurat_obj, config = NULL)
}
\arguments{
\item{seurat_obj}{Seurat object.}

\item{config}{List. Configuration object.}
}
\description{
Quick Summary of Seurat Object for GeoMx Data
}
