% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{Summarizes data by giving count, mean, standard deviation, standard error of the mean, and confidence interval}
\source{
Taken from the R cookbook (cookbook-r.com/Manipulating_data/Summarizing_data/)
}
\usage{
summary_stats(data = NULL, measurevar = NULL, groupvars = NULL,
  na.rm = FALSE, conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{measurevar}{the name of a column that contains the variable to be summariezed}

\item{groupvars}{a vector containing names of columns that contain grouping variables}

\item{na.rm}{a boolean that indicates whether to ignore NAs}

\item{conf.interval}{the percent range of the confidence interval (default is 95%)}

\item{.drop}{Boolean}
}
\description{
Summarizes data by giving count, mean, standard deviation, standard error of the mean, and confidence interval
}
