% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{synthesize_ranks}
\alias{synthesize_ranks}
\title{여러 FindMarkers 결과의 순위를 종합}
\usage{
synthesize_ranks(marker_list, rank_by = "p_val")
}
\arguments{
\item{marker_list}{FindMarkers 결과 데이터프레임의 리스트.}

\item{rank_by}{어떤 컬럼을 기준으로 순위를 매길지 지정. ('p_val' 또는 'avg_log2FC')}
}
\value{
유전자별 종합 순위 정보 데이터프레임.
}
\description{
여러 FindMarkers 결과의 순위를 종합
}
