% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tct.R
\name{tct}
\alias{tct}
\title{Calcula medidas provindas da Teoria Classica dos Testes (TCT)}
\usage{
tct(dados, dadosCorr, mapa, Alts = c("A", "B", "C", "D", "E", "*", "."),
  arred = 10, sdpop = FALSE, summary = FALSE)
}
\arguments{
\item{dados}{Banco de dados a ser utilizado. Este deve ser composto com os seguintes campos: (ID, CAD, PESO, IT1, ..., ITm, Desempenho).}

\item{dadosCorr}{Banco de dados corrigido a ser utilizado. Este deve ser composto com os seguintes campos: (ID, CAD, PESO, IT1, ..., ITm, Desempenho).}

\item{mapa}{DF com informacoes adicionais dos itens. Este deve estar ordenado na mesma ordem que 'dados' e 'dadosCorr', e obrigatoriamente deve conter uma coluna chama 'Gabarito'.}

\item{Alts}{Vetor com as possiveis marcacoes presentes nos dados.}

\item{arred}{Numero de casas decimais a serem consideradas no arredondamento.}

\item{sdpop}{Considerar o desvio populacional para o calculo das estatisticas?}

\item{summary}{Calcula um conjunto de informacoes sobre os dados utilizados nas estatisticas.}
}
\value{
DF com a seguinte composicao: (mapa, Dif, Disc, Abai, Acim, Bis, PercAlts, BiseAlts). Caso summary = TRUE, retorna uma lista com os dois objetos.
}
\description{
Utilizando uma estrutura de BI ou BIB, calcula as estatisticas classicas dos itens.
}
\details{
Etapa Anterior: 'escore'.

Etapa Posterior: 'pos.prop'.
}
\examples{
bib = list(CAD1 = c(1,2), CAD2 = c(2,3), CAD3 = c(3,1))

nblocos = 3

b.len = 2

resp = data.frame(matrix(sample(c("A", "B", "C"), 40, replace = T), ncol = 4), stringsAsFactors = F); names(resp) = paste0("IT", 1:4)

dados = cbind(data.frame(ID = 1:10, CAD = paste0("CAD", sample(3, 10, replace = T)), PESO = 1), resp)

dados = vetor.extendido(dados, bib, nblocos, b.len)

dadosCorr = corrigir(dados, c("A", "B", "A", "C", "A", "B"))

dadosCorr = escore(dadosCorr)

dados$Escore = dadosCorr$Escore

mapa = data.frame(Codigo = 1001:1006, Posicao = 1:6, Gabarito = c("A", "B", "A", "C", "A", "B"))

TCT = tct(dados, dadosCorr, mapa, b.len, nblocos, Alts = c("A", "B", "C"))

}
\author{
Brunno Bittencourt
}
