% Generated by roxygen2 (4.0.1): do not edit by hand
\name{th.recode}
\alias{t.recode}
\alias{th.recode}
\alias{threshold.recode}
\title{Threshold recode}
\usage{
th.recode(x, threshold)

threshold.recode(x, threshold)
}
\arguments{
\item{x}{A number to be recoded}

\item{threshold}{The threshold for recoding}
}
\description{
Recode data based on a threshold. Data >= threshold will be 1, data < threshold = 0.

Threshold recode (alternate name)
}

