% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aggregation.R
\name{utils_aggregation}
\alias{utils_aggregation}
\title{Cell Aggregation Utilities}
\description{
Functions for aggregating cells by various grouping variables
(cluster, sample, patient, etc.) for pseudobulk analysis.
}
