% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{validate_choice}
\alias{validate_choice}
\title{Validate Choice}
\usage{
validate_choice(value, param_name, choices, multiple = FALSE)
}
\arguments{
\item{value}{Value to validate}

\item{param_name}{Name of parameter (for error messages)}

\item{choices}{Vector of allowed values}

\item{multiple}{Whether multiple choices are allowed}
}
\value{
The validated value (or values if multiple=TRUE)
}
\description{
Validates that a value is one of allowed choices (like match.arg but more informative).
}
\keyword{internal}
