% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{validate_genes}
\alias{validate_genes}
\title{Validate Gene List}
\usage{
validate_genes(
  obj,
  genes,
  min_present = NULL,
  assay = NULL,
  warn_missing = TRUE
)
}
\arguments{
\item{obj}{Seurat object or character vector of available genes}

\item{genes}{Character vector of genes to validate}

\item{min_present}{Minimum number of genes that must be present (default: all)}

\item{assay}{Assay to check genes in (for Seurat objects)}

\item{warn_missing}{Whether to warn about missing genes}
}
\value{
Character vector of valid genes present in the object
}
\description{
Checks if genes exist in the object and optionally filters to valid genes.
}
\keyword{internal}
