% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vetor.estendido.R
\name{vetor.estendido}
\alias{vetor.estendido}
\title{Cria um banco de dados em formato estendido}
\usage{
vetor.estendido(dados, bib, nblocos, b.len)
}
\arguments{
\item{dados}{Banco de dados a ser aberto. Este deve ser composto com os seguintes campos: (ID, CAD, PESO, IT1, ..., ITn).}

\item{bib}{Lista com a estrutura de cada caderno. Os nomes de cada vetor dentro da lista deve ser igual ao respectivo valor no campo 'CAD' dentro dos dados.}

\item{nblocos}{Numero de blocos no teste.}

\item{b.len}{Numero de itens em cada bloco.}
}
\value{
DF com a seguinte composicao: (ID, CAD, PESO, IT1, ..., ITn, ..., ITm).
}
\description{
Utilizando uma estrutura de BI ou BIB, esta funcao tem como proposito criar um banco de dados no formato estendido.
}
\details{
Etapa Posterior: 'corrigir'.
}
\examples{
bib = list(CAD1 = (1,2), CAD2 = c(2,3), CAD3 = c(3,1))

nblocos = 3

b.len = 2

resp = data.frame(matrix(sample(c("A", "B", "C"), 40, replace = T), ncol = 4), stringsAsFactors = F); names(resp) = paste0("IT", 1:4)

dados = cbind(data.frame(ID = 1:10, CAD = paste0("CAD", sample(3, 10, replace = T)), PESO = 1), resp)

dados = vetor.estendido(dados, bib, nblocos, b.len)

}
\author{
Brunno Bittencourt
}
