% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demulti_utils.R
\name{generate_sample_names}
\alias{generate_sample_names}
\title{c("a","b","c")과 같이 문자 벡터를 넣으면, "a+b", "b+c" 형태의 조합을 만들어줌.}
\usage{
generate_sample_names(vector)
}
\description{
c("a","b","c")과 같이 문자 벡터를 넣으면, "a+b", "b+c" 형태의 조합을 만들어줌.
}
