% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demulti_utils.R
\name{generate_sample_values}
\alias{generate_sample_values}
\title{1:6과 같이 숫자 벡터를 넣으면, 1+2, 2+3 형태의 조합을 만들어줌.}
\usage{
generate_sample_values(start_num, end_num)
}
\description{
1:6과 같이 숫자 벡터를 넣으면, 1+2, 2+3 형태의 조합을 만들어줌.
}
