% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demulti_utils.R
\name{get_best_two}
\alias{get_best_two}
\title{Get Best Two Probabilities}
\usage{
get_best_two(probs)
}
\arguments{
\item{probs}{Numeric vector of probabilities}
}
\value{
List containing:
\item{best}{Value of the highest probability}
\item{best_idx}{Index of the highest probability}
\item{second}{Value of the second highest probability}
\item{second_idx}{Index of the second highest probability}
}
\description{
Finds the two highest probabilities from a vector.
}
\examples{
get_best_two(c(0.1, 0.7, 0.2))
# Returns: list(best=0.7, best_idx=2, second=0.2, second_idx=3)

}
