% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demulti_utils.R
\name{is_doublet}
\alias{is_doublet}
\title{Check if Sample is Doublet}
\usage{
is_doublet(sample_name)
}
\arguments{
\item{sample_name}{Character vector of sample names}
}
\value{
Logical vector indicating doublet status
}
\description{
Determines if a sample identifier represents a doublet (contains "+").
}
\examples{
is_doublet(c("Sample1", "Sample1+Sample2", "Sample3"))
# Returns: c(FALSE, TRUE, FALSE)

}
