% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{myr_library}
\alias{myr_library}
\title{First line is title}
\usage{
myr_library(package)
}
\description{
This is a brief description.
}
\details{
Anything else after the description goes to the Details section.
}
