% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_working.R
\name{runMUSCAT}
\alias{runMUSCAT}
\title{Muscat (Pseudo-bulking) 파이프라인 함수 (V5 - prepData 사용)}
\usage{
runMUSCAT(
  sobj,
  cluster_id = "seurat_clusters",
  sample_id = "hos_no",
  group_id = "type",
  batch_id = NULL,
  contrast = NULL,
  method = "edgeR",
  pb_min_cells = 3,
  filter_genes = c("none", "genes", "both", "edgeR"),
  keep_clusters = NULL,
  cluster_label_map = NULL
)
}
\arguments{
\item{sobj}{(Seurat) Seurat 객체}

\item{cluster_id}{(character) 세포 유형 컬럼명}

\item{sample_id}{(character) 환자/샘플 ID 컬럼명}

\item{group_id}{(character) 비교할 그룹 컬럼명}

\item{contrast}{(character) Contrast (예: "groupA-groupB")}

\item{method}{(character) "edgeR", "DESeq2", "limma-trend", "limma-voom" (block 없음)}

\item{formula_str}{(character) 포뮬러 (예: "~ 0 + group + set")
                'group' 키워드는 'group_id'로 자동 변환됩니다.}
}
\description{
prepData를 사용하여 SCE 객체를 포맷하고, pbDS로 GLM(edgeR/DESeq2) 실행
}
\note{
이 함수는 'pbDS'의 한계로 인해 혼합 효과(Random Effects/block)를 지원하지 않습니다.
}
